Allow access to Location short names
This project only provided access to the long version of each location
variable. For example, only the full US state name (e.g. Wyoming)
could be provided. It was not possible to get the short version of
the name (e.g. WY).

This change allows access to the short Locality, Town, Country,
District, and Street Address.